#54 Add involvement check before updating/deleting offers  (#61)
* add author check for delete&update offer

* move duplicated code to separate method

+++ src/services/offer.js
@@ -1,4 +1,5 @@
 const Offer = require('~/models/offer')
+const { createForbiddenError } = require('~/utils/errorsHelper')
 
 const filterAllowedFields = require('~/utils/filterAllowedFields')
 const { allowedOfferFieldsForUpdate } = require('~/validation/services/offer')
@@ -50,10 +51,9 @@ const offerService = {
   },
 
   updateOffer: async (id, currentUserId, updateData) => {
+    await offerService.checkOfferAuthor(id, currentUserId)
     const filteredUpdateData = filterAllowedFields(updateData, allowedOfferFieldsForUpdate)
-
     const offer = await Offer.findById(id)
-
     for (let field in filteredUpdateData) {
       offer[field] = filteredUpdateData[field]
     }
@@ -62,8 +62,18 @@ const offerService = {
     await offer.save()
   },
 
-  deleteOffer: async (id) => {
+  deleteOffer: async (id, currentUserId) => {
+    await offerService.checkOfferAuthor(id, currentUserId)
     await Offer.findByIdAndRemove(id).exec()
+  },
+
+  checkOfferAuthor: async (id, currentUserId) => {
+    const offer = await Offer.findById(id)
+    const author = offer.author.toString()
+
+    if (currentUserId !== author) {
+      throw createForbiddenError()
+    }
   }
 }
